Extract ConditionalExpression above Comparison and IN/LIKE
