+++ component/component-commons/src/main/java/org/bithon/component/commons/expression/BinaryExpression.java
@@ -16,8 +16,6 @@
 
 package org.bithon.component.commons.expression;
 
-import java.util.Set;
-
 /**
  * @author frank.chen021@outlook.com
  * @date 2023/4/7 20:17
@@ -56,131 +54,5 @@ public abstract class BinaryExpression implements IExpression {
     @Override
     public String getType() {
         return type;
-    }
-
-    @Override
-    public void accept(IExpressionVisitor visitor) {
-        if (visitor.visit(this)) {
-            this.left.accept(visitor);
-            this.right.accept(visitor);
-        }
-    }
-
-    @Override
-    public <T> T accept(IExpressionVisitor2<T> visitor) {
-        return visitor.visit(this);
-    }
-
-    public static class In extends BinaryExpression {
-
-        public In(IExpression left, ExpressionList right) {
-            this("in", left, right);
-        }
-
-        protected In(String operator, IExpression left, ExpressionList right) {
-            super(operator, left, right);
-        }
-
-        @Override
-        public IDataType getDataType() {
-            return IDataType.BOOLEAN;
-        }
-
-        @SuppressWarnings("unchecked")
-        @Override
-        public Object evaluate(IEvaluationContext context) {
-            Object l = left.evaluate(context);
-            if (l == null) {
-                return false;
-            }
-
-            Set<Object> sets = (Set<Object>) right.evaluate(context);
-            Object o = sets.iterator().next();
-            if (o instanceof String) {
-                return sets.contains(l.toString());
-            }
-            if (o instanceof Long) {
-                return sets.contains(toLong(l));
-            }
-            if (o instanceof Double) {
-                return sets.contains(toDouble(l));
-            }
-
-            throw new UnsupportedOperationException("Type of " + o.getClass().getSimpleName() + " is not supported by IN operator");
-        }
-
-        private Object toLong(Object o) {
-            if (o instanceof Long) {
-                return o;
-            }
-            if (o instanceof Number) {
-                return ((Number) o).longValue();
-            }
-            return Long.parseLong(o.toString());
-        }
-
-        private Object toDouble(Object o) {
-            if (o instanceof Double) {
-                return o;
-            }
-            if (o instanceof Number) {
-                return ((Number) o).doubleValue();
-            }
-            return Double.parseDouble(o.toString());
-        }
-    }
-
-    public static class Like extends BinaryExpression {
-
-        public Like(IExpression left, IExpression right) {
-            this("like", left, right);
-        }
-
-        public Like(String operator, IExpression left, IExpression right) {
-            super(operator, left, right);
-        }
-
-        @Override
-        public IDataType getDataType() {
-            return IDataType.BOOLEAN;
-        }
-
-        @Override
-        public Object evaluate(IEvaluationContext context) {
-            String r = (String) right.evaluate(context);
-
-            // Escape any special characters in the pattern
-            String pattern = r.replaceAll("%", "\\\\%").replaceAll("_", "\\\\_");
-
-            // Replace SQL wildcard characters with Java regex wildcard characters
-            pattern = pattern.replaceAll("%", ".*").replaceAll("_", ".");
-
-            String l = (String) left.evaluate(context);
-            return l != null && l.contains(pattern);
-        }
-    }
-
-    public static class NotLike extends Like {
-
-        public NotLike(IExpression left, IExpression right) {
-            super("not like", left, right);
-        }
-
-        @Override
-        public Object evaluate(IEvaluationContext context) {
-            return !((boolean) super.evaluate(context));
-        }
-    }
-
-    public static class NotIn extends In {
-
-        public NotIn(IExpression left, ExpressionList right) {
-            super("not in", left, right);
-        }
-
-        @Override
-        public Object evaluate(IEvaluationContext context) {
-            return !((boolean) super.evaluate(context));
-        }
     }
 }
