+++ component/component-commons/src/main/java/org/bithon/component/commons/expression/IExpressionVisitor2.java
@@ -49,11 +49,7 @@ public interface IExpressionVisitor2<T> {
         return null;
     }
 
-    default T visit(BinaryExpression expression) { return null; }
-
-    default T visit(ComparisonExpression expression) {
-        return null;
-    }
+    default T visit(ConditionalExpression expression) { return null; }
 
     default T visit(MacroExpression expression) {
         return null;
