+++ server/storage-jdbc-mysql/src/main/java/org/bithon/server/storage/jdbc/mysql/MySQLSqlDialect.java
@@ -17,7 +17,7 @@
 package org.bithon.server.storage.jdbc.mysql;
 
 import com.fasterxml.jackson.annotation.JsonTypeName;
-import org.bithon.component.commons.expression.BinaryExpression;
+import org.bithon.component.commons.expression.ConditionalExpression;
 import org.bithon.component.commons.expression.FunctionExpression;
 import org.bithon.component.commons.expression.IExpression;
 import org.bithon.component.commons.expression.IdentifierExpression;
@@ -100,8 +100,8 @@ public class MySQLSqlDialect implements ISqlDialect {
                         patternExpression = new FunctionExpression(Functions.getInstance().getFunction("concat"),
                                                                    Arrays.asList(patternExpression, LiteralExpression.create("%")));
                     }
-                    return new BinaryExpression.Like(expression.getParameters().get(0),
-                                                     patternExpression);
+                    return new ConditionalExpression.Like(expression.getParameters().get(0),
+                                                          patternExpression);
                 } else if ("endsWith".equals(expression.getName())) {
                     // MySQL does not provide endsWith function, turns it into LIKE expression as: LIKE '%prefix'
                     IExpression patternExpression = expression.getParameters().get(1);
@@ -111,8 +111,8 @@ public class MySQLSqlDialect implements ISqlDialect {
                         patternExpression = new FunctionExpression(Functions.getInstance().getFunction("concat"),
                                                                    Arrays.asList(LiteralExpression.create("%"), patternExpression));
                     }
-                    return new BinaryExpression.Like(expression.getParameters().get(0),
-                                                     patternExpression);
+                    return new ConditionalExpression.Like(expression.getParameters().get(0),
+                                                          patternExpression);
                 }
 
                 return expression;
